feat: add blocks toolbar (#2272)
