+++ apps/journeys-admin/src/components/Editor/Canvas/AddBlockToolbar/NewImageButton/NewImageButton.spec.tsx
