+++ apps/journeys-admin/src/components/Editor/Canvas/AddBlockToolbar/NewImageButton/index.ts
