+++ apps/journeys-admin/src/components/Editor/Canvas/AddBlockToolbar/NewSignUpButton/NewSignUpButton.tsx
@@ -11,7 +11,7 @@ import Mail2Icon from '@core/shared/ui/icons/Mail2'
 
 import { GetJourney_journey_blocks_CardBlock as CardBlock } from '../../../../../../__generated__/GetJourney'
 import { SignUpBlockCreate } from '../../../../../../__generated__/SignUpBlockCreate'
-import { Button } from '../../Button'
+import { Button } from '../Button'
 
 export const SIGN_UP_BLOCK_CREATE = gql`
   ${SIGN_UP_FIELDS}
