+++ apps/journeys-admin/src/components/Editor/Canvas/AddBlockToolbar/NewTextResponseButton/NewTextResponseButton.tsx
@@ -11,7 +11,7 @@ import TextInput1Icon from '@core/shared/ui/icons/TextInput1'
 
 import { GetJourney_journey_blocks_CardBlock as CardBlock } from '../../../../../../__generated__/GetJourney'
 import { TextResponseBlockCreate } from '../../../../../../__generated__/TextResponseBlockCreate'
-import { Button } from '../../Button'
+import { Button } from '../Button'
 
 export const TEXT_RESPONSE_BLOCK_CREATE = gql`
   ${TEXT_RESPONSE_FIELDS}
