+++ apps/journeys-admin/src/components/Editor/Canvas/AddBlockToolbar/index.ts
@@ -1,0 +1,1 @@
+export { AddBlockToolbar } from './AddBlockToolbar'
