+++ apps/journeys-admin/src/components/Editor/Canvas/Canvas.tsx
@@ -29,6 +29,11 @@ import { QuickControls } from './QuickControls'
 import { SelectableWrapper } from './SelectableWrapper'
 import { VideoWrapper } from './VideoWrapper'
 import { ActiveSlide } from '@core/journeys/ui/EditorProvider/EditorProvider'
+import { AddBlockToolbar } from './AddBlockToolbar'
+import Card from '@mui/material/Card'
+
+import { GetJourney_journey_blocks_CardBlock as CardBlock } from '../../../../__generated__/GetJourney'
+import { TreeBlock } from '@core/journeys/ui/block'
 
 const NextCard = dynamic(
   async () =>
@@ -133,6 +138,10 @@ export function Canvas(): ReactElement {
     })
   }
 
+  const cardBlock = selectedStep?.children.find(
+    (block) => block.__typename === 'CardBlock'
+  ) as TreeBlock<CardBlock>
+
   return (
     <Stack
       onClick={handleSelectCard}
@@ -147,7 +156,7 @@ export function Canvas(): ReactElement {
       <Box
         sx={{
           width: activeSlide === ActiveSlide.Canvas ? 50 : 0,
-          mr: activeSlide === ActiveSlide.Canvas ? 4 : 0,
+          mr: activeSlide === ActiveSlide.Canvas ? 7 : 0,
           transition: (theme) =>
             theme.transitions.create(['width', 'margin'], {
               duration: 150
@@ -259,15 +268,23 @@ export function Canvas(): ReactElement {
       <Box
         sx={{
           width: activeSlide === ActiveSlide.Canvas ? 50 : 0,
-          ml: activeSlide === ActiveSlide.Canvas ? 4 : 0,
+          ml: activeSlide === ActiveSlide.Canvas ? 7 : 0,
           transition: (theme) =>
             theme.transitions.create(['width', 'margin'], {
               duration: 150
             }),
-          overflow: 'hidden'
+          overflow: 'hidden',
+          alignSelf: 'center'
         }}
       >
-        {/* ADD BLOCK COMPONENT GOES HERE */}
+        <Card
+          variant="outlined"
+          sx={{
+            borderRadius: 2
+          }}
+        >
+          <AddBlockToolbar selectedCard={cardBlock} />
+        </Card>
       </Box>
     </Stack>
   )
