--- apps/journeys-admin/src/components/Editor/ControlPanel/BlocksTab/index.ts
@@ -1,1 +1,0 @@
-export { BlocksTab } from './BlocksTab'
